+++ R/train_lightgbm.R
@@ -1,30 +1,19 @@
-train_lightgbm = function(self, task, task_type) {
+train_lightgbm = function(self, task, task_type, pars, init_model = NULL) {
   assert_choice(task_type, c("regr", "classif"))
-
   self$state$feature_names = task$feature_names
-  pars = self$param_set$get_values(tags = "train")
-
   convert_categorical = pars$convert_categorical
   pars$convert_categorical = NULL
-
-  if (!is.null(pars$custom_eval)) {
-    pars$metric = pars$custom_eval
-    pars$custom_eval = NULL
-  }
-
-  early_stopping_split = pars$early_stopping_split
-  pars$early_stopping_split = NULL
 
   if (convert_categorical) {
     encoding = encode_lightgbm_train(task)
     X = encoding$X
-    categorical_feature = encoding$categorical_feature
+    categorical_feature = c(encoding$categorical_feature, pars$categorical_feature)
   } else {
     assert_true(all(task$feature_types$type %in% c("integer", "numeric")))
     X = data.matrix(task$data(cols = task$feature_names))
     categorical_feature = pars$categorical_feature
-    pars$categorical_feature = NULL
   }
+  pars$categorical_feature = NULL
 
   y = task$data(cols = task$target_names)[[1L]]
 
@@ -58,8 +47,11 @@ train_lightgbm = function(self, task, task_type) {
 
   }
 
-  if (!is.null(early_stopping_split) && early_stopping_split > 0) {
-    # we cannot simultaneously to stratification and grouping
+  early_stopping_split = pars$early_stopping_split
+  pars$early_stopping_split = NULL
+
+  if (isTRUE(early_stopping_split > 0)) {
+    # we cannot simultaneously do stratification and grouping
     ids = mlr3::partition(task, ratio = 1 - early_stopping_split, stratify = is.null(task$groups))
 
     self$state$valid_ids = ids$test
@@ -83,8 +75,8 @@ train_lightgbm = function(self, task, task_type) {
 
     row_id = NULL
     if ("weights" %in% task$properties) {
-      dtrain$setinfo("weight", subset(task$weights, row_id %in% ids$train)$weight)
-      dvalid$setinfo("weight", subset(task$weights, row_id %in% ids$test)$weight)
+      dtrain$set_field("weight", subset(task$weights, row_id %in% ids$train)$weight)
+      dvalid$set_field("weight", subset(task$weights, row_id %in% ids$test)$weight)
     }
   } else {
     dtrain = lightgbm::lgb.Dataset(
@@ -96,20 +88,21 @@ train_lightgbm = function(self, task, task_type) {
     valids = list()
 
     if ("weights" %in% task$properties) {
-      dtrain$setinfo("weight", task$weights$weight)
+      dtrain$set_field("weight", task$weights$weight)
     }
   }
 
-  args = pars[which(names(pars) %in% formalArgs(lightgbm::lgb.train))]
-  params = pars[which(names(pars) %nin% formalArgs(lightgbm::lgb.train))]
+  ii = names(pars) %in% formalArgs(lightgbm::lgb.train)
+  args = pars[ii]
+  params = pars[!ii]
 
-  invoke(
+ invoke(
     lightgbm::lgb.train,
     data = dtrain,
     valids = valids,
     .args = args,
     params = params,
-    categorical_feature = categorical_feature
+    categorical_feature = categorical_feature,
+    init_model = init_model
   )
-
 }